Separate firebase from main.ts
